Strip metadata from notebooks to play nice in git
While the notebook is somewhat friendly to text editors, it keeps
copies of all output, including binary blobs for images.  This
means that there will be a lot of diff noise, so added a filter in
my global gitconfig and created a .gitattributes file to use the
filter for all ipynb files.

It uses `jq` to strip all the extra data.